refactor(client): extract upload helpers from chat_input_bar.dart
Move four shared upload constants/pure-Dart helpers out of the
2030-LOC chat_input_bar.dart god module into a dedicated file under
the existing chat_input_bar/ subdirectory:

- kOctetStream (default MIME subtype fallback)
- kMaxUploadBytes (mirrors server's MAX_FILE_SIZE)
- genericFilename(String) (used by the privacy 'preserve original
  filenames' toggle)
- formatBytes(int) (human-readable byte counts)

These were used by three separate picker code paths in the file
(_pickFile, _pickImageFromGallery, _pickImageFromCamera) plus the
clipboard-paste flow. Renamed from leading-underscore private to
public symbols so they're importable; six call sites updated.

Honest scope note: the rest of chat_input_bar.dart resists mechanical
extraction — every section (mentions, drafts, voice, edit mode,
attachment helpers, build helpers) references private state of the
parent State class, so further splits would require either making
fields public or factoring the state into a separate object. That's
its own focused PR. Net delta here: -27 LOC from the parent file,
+44 LOC in the new helpers file (the latter is bigger because it
adds module-level documentation and a library directive). 1332 tests
pass.
